+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [1.1.1] - 2026-04-26
+
+- add (Plus) mode, where additional separators are supported.
+- original "SWAP: Head and Tail" command now simply supports whitespace separators.
+
 ## [1.1.0] - 2026-04-05
 
 - Support more separators, including `,`, `;`, `|`, `"`, `<`, `>`, `=`, `<=`, `>=`, `!=`.